Module tame_gcs::signing

Helper facilities for calculating content digests and signing data

Structs

ServiceAccount
A GCP service account, used as a KeyProvider when signing URLs.

Enums

DigestAlgorithm
The supported algorithms for creating a digest of content

Key
The supported key formats

SigningAlgorithm
The supported algorithms for signing payloads

Traits

DigestCalulator
Used to calculate a digest of payloads with a specific algorithm

KeyProvider
Provides the details needed for signing a URL

Signer
Used to sign a block of data