Trait tame_gcs::signing::DigestCalulator

pub trait DigestCalulator {
    fn digest(
        &self, 
        algorithm: DigestAlgorithm, 
        data: &[u8], 
        output_digest: &mut [u8]
    );
}

Used to calculate a digest of payloads with a specific algorithm

Required methods

fn digest(
&self,
algorithm: DigestAlgorithm,
data: &[u8],
output_digest: &mut [u8]
)

Calculate a digest of a block of data, the algorithm determines the size of the slice used for returning the digest

Loading content...